Issue #1: 초기 이관 및 RustBridge 통합 정리

+++ Sources/RhwpCoreBridge/RhwpDocument.swift
@@ -44,7 +44,7 @@ class RhwpDocument {
             guard let base = rawBuffer.baseAddress else { return nil }
             return rhwp_open(
                 base.assumingMemoryBound(to: UInt8.self),
-                data.count
+                UInt(data.count)
             )
         }
         guard let validHandle = result else {
